kr3/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
PyJWT
python-dotenv
slowapi
